RDS/circle1_adapters_and_ports/port_zenodo/requirements.txt
-i https://pypi.org/simple
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
click==7.0
clickclick==1.2.2
connexion-plus==0.14
connexion[swagger-ui]==2.4.0
contextlib2==0.6.0.post1
flask-cors==3.0.8
flask-opentracing==1.1.0
flask==1.1.1
future==0.18.2
gevent==1.4.0
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itsdangerous==1.1.0
jaeger-client==4.1.0
jinja2==2.10.3
jsonschema==3.1.1
markdown==3.1.1
markupsafe==1.1.1
more-itertools==7.2.0
openapi-spec-validator==0.2.8
opentracing-instrumentation==3.2.1
opentracing==2.2.0
prometheus-client==0.7.1
prometheus-flask-exporter==0.11.0
pyrsistent==0.15.5
python-jsonschema-objects==0.3.11
pyyaml==5.1.2
requests==2.22.0
six==1.13.0
swagger-ui-bundle==0.0.6
threadloop==1.0.2
thrift==0.11.0
tornado==5.1.1
urllib3==1.25.6
werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
